+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"27.05.2026 09:55"
-#define NE_VERSION   L"2026.05.27.09"
+#define NE_PUBLISHED L"27.05.2026 10:14"
+#define NE_VERSION   L"2026.05.27.10"